self-hosted/requirements.txt
# Core
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Slack integration (optional)
slack-bolt>=1.18.0
slack-sdk>=3.21.0

# Screenshot support (optional)
playwright>=1.40.0

# Subprocess interaction
pexpect>=4.8.0
